+++ packages/site/examples/item/customCombo/demo/cCircle.js
@@ -44,6 +44,7 @@ G6.registerCombo(
           r: style.r,
         },
         draggable: true,
+        // must be assigned in G6 3.3 and later versions. it can be any string you want, but should be unique in a custom item type
         name: 'combo-keyShape',
       });
       // Add the marker on the bottom
@@ -58,6 +59,7 @@ G6.registerCombo(
           symbol: collapseIcon,
         },
         draggable: true,
+        // must be assigned in G6 3.3 and later versions. it can be any string you want, but should be unique in a custom item type
         name: 'combo-marker-shape',
       });